backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
requests==2.31.0
waitress==3.0.2
werkzeug==2.3.7
Pillow